Peak content

+++ server.js
@@ -15,8 +15,50 @@ let context = ""
 const dirArr = []
 const fileArr = []
 
+function dirExists(filepath, dirIndex) {
+    if (dirIndex == 0) {
+        dirIndex += 1
+        filepath = `${filepath} (${dirIndex})`
+    } else {
+        dirIndex += 1
+        filepath = filepath.slice(0, -4)
+        filepath = `${filepath} (${dirIndex})`
+    }
+    try {
+        fs.mkdirSync(filepath)
+    }
+    catch (err) {
+        console.log(err.code)
+        if (err.code == 'EEXIST') {
+            console.log(dirIndex)
+            dirExists(filepath, dirIndex)
+        }
+    }
+}
+
+function fileExists(filepath, dirIndex) {
+    if (dirIndex == 0) {
+        dirIndex += 1
+        filepath = `${filepath} (${dirIndex})`
+    } else {
+        dirIndex += 1
+        filepath = filepath.slice(0, -4)
+        filepath = `${filepath} (${dirIndex})`
+    }
+    try {
+        fs.mkdirSync(filepath)
+    }
+    catch (err) {
+        console.log(err.code)
+        if (err.code == 'EEXIST') {
+            console.log(dirIndex)
+            dirExists(filepath, dirIndex)
+        }
+    }
+}
+
 // GET-y
-app.get("/filemanager", function (req, res) {
+app.get("/", function (req, res) {
     fs.readdir(`${__dirname}\\upload`, (err, files) => {
         if (err) throw err
         console.log("lista", files);
@@ -41,15 +83,19 @@ app.get("/filemanager", function (req, res) {
 })
 
 app.post("/addFolder", function (req, res) {
+    let dirIndex = 0;
     console.log('addFolder')
     console.log(req.body)
-    const filepath = path.join(__dirname, "upload", req.body.name)
+    let filepath = path.join(__dirname, "upload", req.body.name)
     console.log(filepath)
     try {
         fs.mkdirSync(filepath)
     }
     catch (err) {
-        console.log("error", err.message);
+        console.log(err.code)
+        if (err.code == 'EEXIST') {
+            dirExists(filepath, dirIndex)
+        }
     }
     res.render('index.hbs', context)
 })
@@ -59,10 +105,15 @@ app.post("/addFile", function (req, res) {
     console.log(req.body)
     const filepath = path.join(__dirname, "upload", req.body.name)
     console.log(filepath)
-    fs.writeFile(filepath, '', (err) => {
-        if (err) throw err
-        console.log("plik nadpisany");
-    })
+    if (fs.existsSync(filepath)) {
+        console.log("plik istnieje");
+    } else {
+        fs.writeFile(filepath, '', (err) => {
+            if (err) throw err
+            console.log("plik nadpisany");
+        })
+    }
+
     res.render('index.hbs', context)
 })
